+++ scripts/buildings/types/Fishingpost.cs
@@ -5,7 +5,7 @@ using System;
 public partial class Fishingpost : BuildingBase
 {
     public override int MaxBuilders => 3;
-    public override double TotalBuildingProgressNeeded => 30;
+    public override float TotalBuildingProgressNeeded => 30;
     public override bool IsResourceStorage => false;
     public override BuildingType Type => BuildingType.FishingPost;
     // Called when the node enters the scene tree for the first time.
