+++ scripts/buildings/types/Lumberjack.cs
@@ -12,7 +12,7 @@ namespace SacaSimulationGame.scripts.buildings
     public partial class Lumberjack : StorageBuildingBase
     {
         public override int MaxBuilders => 2;
-        public override double TotalBuildingProgressNeeded => 5;
+        public override float TotalBuildingProgressNeeded => 5;
         public override bool IsResourceStorage => false;
     
         public override BuildingType Type => BuildingType.Lumberjack;
